requirements.txt
discord.py>=2.3.2
web3==6.13.0
websockets
PyNaCl
PyNaCl